refactor: NoSuchResourceException 클래스들 service 패키지로 이동
