+++ src/main/java/subway/adapter/in/web/exception/ExceptionAdvice.java
@@ -13,8 +13,8 @@ import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.MissingServletRequestParameterException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
+import subway.application.service.exception.SubwayNoSuchResourceException;
 import subway.common.exception.SubwayIllegalArgumentException;
-import subway.common.exception.SubwayNoSuchResourceException;
 
 @RestControllerAdvice
 public class ExceptionAdvice {
