+++ src/main/java/subway/application/service/section/SectionCommandService.java
@@ -9,8 +9,8 @@ import subway.application.port.in.section.dto.command.RemoveStationFromLineComma
 import subway.application.port.out.line.LoadLinePort;
 import subway.application.port.out.line.PersistLinePort;
 import subway.application.port.out.station.LoadStationPort;
-import subway.common.exception.NoSuchLineException;
-import subway.common.exception.NoSuchStationException;
+import subway.application.service.exception.NoSuchLineException;
+import subway.application.service.exception.NoSuchStationException;
 import subway.domain.line.Line;
 import subway.domain.section.Section;
 import subway.domain.station.Station;
