+++ src/main/java/subway/application/service/station/StationCommandService.java
@@ -13,7 +13,7 @@ import subway.application.port.in.station.dto.command.UpdateStationCommand;
 import subway.application.port.out.line.LoadLinePort;
 import subway.application.port.out.station.LoadStationPort;
 import subway.application.port.out.station.PersistStationPort;
-import subway.common.exception.NoSuchStationException;
+import subway.application.service.exception.NoSuchStationException;
 import subway.common.exception.SubwayIllegalArgumentException;
 import subway.domain.station.Station;
