+++ src/test/java/subway/application/service/line/LineCommandServiceTest.java
@@ -15,7 +15,7 @@ import subway.application.port.in.line.dto.command.CreateLineCommand;
 import subway.application.port.in.line.dto.command.UpdateLineInfoCommand;
 import subway.application.port.out.line.LoadLinePort;
 import subway.application.port.out.line.PersistLinePort;
-import subway.common.exception.NoSuchLineException;
+import subway.application.service.exception.NoSuchLineException;
 import subway.common.exception.SubwayIllegalArgumentException;
 
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
