+++ src/test/java/subway/application/service/route/RouteQueryServiceTest.java
@@ -19,14 +19,14 @@ import subway.application.port.in.route.dto.response.RouteQueryResponse;
 import subway.application.port.out.line.LoadLinePort;
 import subway.application.port.out.route.RouteFinderPort;
 import subway.application.port.out.station.LoadStationPort;
-import subway.common.exception.NoSuchStationException;
-import subway.domain.line.Line;
-import subway.domain.section.Section;
-import subway.domain.station.Station;
+import subway.application.service.exception.NoSuchStationException;
 import subway.domain.fare.Fare;
 import subway.domain.fare.FarePolicy;
+import subway.domain.line.Line;
 import subway.domain.route.Route;
 import subway.domain.route.RouteSection;
+import subway.domain.section.Section;
+import subway.domain.station.Station;
 import subway.fixture.StationFixture.강남역;
 import subway.fixture.StationFixture.삼성역;
