+++ src/test/java/subway/application/service/station/StationCommandServiceTest.java
@@ -20,7 +20,7 @@ import subway.application.port.in.station.dto.command.UpdateStationCommand;
 import subway.application.port.out.line.LoadLinePort;
 import subway.application.port.out.station.LoadStationPort;
 import subway.application.port.out.station.PersistStationPort;
-import subway.common.exception.NoSuchStationException;
+import subway.application.service.exception.NoSuchStationException;
 import subway.common.exception.SubwayIllegalArgumentException;
 import subway.fixture.StationFixture.역삼역;
